+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `HistoryBufferView`, the `!Sized` version of `HistoryBuffer`.
 - Added `DequeView`, the `!Sized` version of `Deque`.
 - Added `QueueView`, the `!Sized` version of `Queue`.
+- Added `SortedLinkedListView`, the `!Sized` version of `SortedLinkedList`.
 
 ### Changed